+++ api/src/shared/database/database.py
@@ -11,7 +11,7 @@ from shared.database_gen.sqlacodegen_models import Base, Feed, Gtfsfeed, Gtfsrea
 from sqlalchemy.orm import sessionmaker
 import logging
 
-from utils.logger import get_env_logging_level
+from shared.common.logging_utils import get_env_logging_level
 
 
 def generate_unique_id() -> str:
