+++ functions-python/batch_process_dataset/function_config.json
@@ -6,7 +6,7 @@
   "memory": "2Gi",
   "trigger_http": true,
   "include_folders": ["helpers", "dataset_service"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "secret_environment_variables": [
     {
       "key": "FEEDS_DATABASE_URL"
