+++ functions-python/operations_api/function_config.json
@@ -6,7 +6,7 @@
   "memory": "1Gi",
   "trigger_http": true,
   "include_folders": ["helpers"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "environment_variables": [
     {
       "key": "GOOGLE_CLIENT_ID"
