+++ functions-python/preprocessed_analytics/function_config.json
@@ -6,7 +6,7 @@
   "memory": "4Gi",
   "trigger_http": false,
   "include_folders": ["helpers"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "environment_variables": [],
   "secret_environment_variables": [
     {
