+++ functions-python/process_validation_report/function_config.json
@@ -6,7 +6,7 @@
   "memory": "2Gi",
   "trigger_http": true,
   "include_folders": ["helpers"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "secret_environment_variables": [
     {
       "key": "FEEDS_DATABASE_URL"
