+++ functions-python/reverse_geolocation/function_config.json
@@ -6,7 +6,7 @@
   "available_memory": "4Gi",
   "trigger_http": true,
   "include_folders": ["helpers"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "environment_variables": [],
   "secret_environment_variables": [
     {
