+++ functions-python/reverse_geolocation_populate/function_config.json
@@ -5,7 +5,7 @@
   "timeout": 3600,
   "trigger_http": true,
   "include_folders": ["helpers"],
-  "include_api_folders": ["database_gen", "database"],
+  "include_api_folders": ["database_gen", "database", "common"],
   "environment_variables": [],
   "secret_environment_variables": [
     {
